requirements.txt
faiss_cpu==1.8.0.post1
sentence_transformers==3.0.1
parameters-validation==1.2.0